requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.2.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0

